FIX: Sharing to Discord forum channels requires a thread_name (#251)
Discourse is able to share topics and replies to normal Discord channels, but it doesn't work with forum channels, which require an additional `thread_name` parameter.

This change fixes the issue by checking for the specific error the Discord returns when trying to share to a forum channel, adding the `thread_name`, then trying again.

This is intentionally a minimal fix, additional work would be required to make the sharing more closely fit the style of Discord forum channels.